requirements.txt
huggingface_hub[cli]
ipykernel
numpy==2.2.4
pillow
pydrive==1.3.1
pymilvus==2.5.5
pymupdf
python-dotenv
transformers